requirements.txt
click==8.0.1
PyYAML==5.4.1
requests==2.26.0
